Improve ValueFormatter precision for small values (#904)
* Improve ValueFormatter precision for small values

Introduces a new formatterMiddle with 4 decimal places for values between 0.01 and 1 in ValueFormatter's auto style. Updates SwapSceneViewModel to use the .auto style and adjusts related tests to reflect the new formatting behavior.

* Adjust ValueFormatter significant digits and tests

Updated ValueFormatter to use significant digits with a minimum of 2 and a maximum of 4, and enabled usesSignificantDigits. Modified the formatter selection logic and updated related unit tests to match the new formatting behavior.

---------

Co-authored-by: gemcoder21 <104884878+gemcoder21@users.noreply.github.com>